+++ src/main/java/apply_to_group_screens/ApplyToGroupFailed.java
@@ -1,4 +1,4 @@
-package ApplyToGroup_screens;
+package apply_to_group_screens;
 
 public class ApplyToGroupFailed extends RuntimeException{
     public ApplyToGroupFailed(String error){
