+++ src/main/java/apply_to_group_use_case/ApplyToGroupInputBoundary.java
@@ -1,4 +1,4 @@
-package ApplyToGroup;
+package apply_to_group_use_case;
 
 public interface ApplyToGroupInputBoundary {
     ApplyToGroupResponseModel applyToGroup(ApplyToGroupRequestModel requestModel);
