+++ src/main/java/apply_to_group_use_case/ApplyToGroupOutputBoundary.java
@@ -1,4 +1,4 @@
-package ApplyToGroup;
+package apply_to_group_use_case;
 
 public interface ApplyToGroupOutputBoundary {
     ApplyToGroupResponseModel prepareSuccessView(ApplyToGroupResponseModel user);
