+++ src/test/java/apply_to_group_use_case/ApplyToGroupInteractorTest.java
@@ -1,4 +1,4 @@
-package ApplyToGroup_use_case;
+package apply_to_group_use_case;
 
 public class ApplyToGroupInteractorTest {
